+++ pyxel/detectors/__init__.py
@@ -12,6 +12,7 @@
 from .channels import Channels, Matrix, ReadoutPosition
 from .readout_properties import ReadoutProperties
 from .environment import Environment, WavelengthHandling
+from .charge_to_volt_settings import ChargeToVoltSettings
 from .detector import Detector
 from .geometry import Geometry
 from .characteristics import Characteristics