+++ pyxel/detectors/characteristics.py
@@ -12,16 +12,118 @@ intrinsic physical and electronic parameters of a detector. These parameters
 include quantum efficiency, charge-to-voltage conversion, pre-amplification,
 full-well capacity, ADC resolution, and voltage range.
 """
-
+import warnings
 from collections.abc import Iterable, Mapping, Sequence
 from typing import TYPE_CHECKING
 
 import numpy as np
+from typing_extensions import Self
 
+from pyxel.detectors import ChargeToVoltSettings
 from pyxel.util import get_size, get_uninitialized_error
 
 if TYPE_CHECKING:
     from pyxel.detectors import Geometry
+
+
+def validate_pre_amplification(
+    pre_amplification: int | float | dict[str, int | float] | None,
+) -> None:
+    """Validate the pre-amplification gain value(s).
+
+    Ensure that the pre-amplification gain(s) are within a valid range.
+
+    Parameters
+    ----------
+    pre_amplification : int, float, dict or None
+        Pre-amplifier gain(s). Can be a single value applied to all pixels
+        or a dictionary mapping channel names to individual gains. Unit: V/V
+
+    Raises
+    ------
+    ValueError
+        If one or more gain values are outside the valid range.
+    TypeError
+        If the provided type is unsupported.
+    """
+    if pre_amplification is None:
+        return
+
+    elif isinstance(pre_amplification, int | float):
+        if not (0.0 <= pre_amplification <= 10_000.0):
+            raise ValueError("'pre_amplification' must be between 0.0 and 10000.0.")
+
+    elif isinstance(pre_amplification, dict):
+        for channel, gain in pre_amplification.items():
+            if not (0.0 <= gain <= 10_000.0):
+                raise ValueError(
+                    f"'pre_amplification' must be between 0.0 and 10000.0. for {channel=}"
+                )
+    else:
+        raise TypeError(
+            "'pre_amplification' must be a number, a dictionary of numbers or None"
+        )
+
+
+def to_pre_amplification_map(
+    pre_amplification: int | float | dict[str, float] | None,
+    geometry: "Geometry",
+) -> float | np.ndarray | None:
+    """Resolve pre-amplification configuration into a number or a 2D array.
+
+    Convert a per-channel pre-amplification gain definition into a 2D array
+    matching the detector geometry.
+    This allows subsequent models to operate channel-wise on pre-amplification values.
+
+    Parameters
+    ----------
+    pre_amplification : int, float, dict or None
+        Pre-amplifier gain(s). Can be a single value applied to all pixels
+        or a dictionary mapping channel names to individual gains. Unit: V/V
+
+    geometry : Geometry
+        Detector geometry definition, including channel layout.
+
+    Returns
+    -------
+    float or ndarray or None
+
+    Raises
+    ------
+    ValueError
+        If the provided channels do not match those in ``geometry``.
+    """
+    # Check
+    if pre_amplification is None:
+        return None
+
+    elif isinstance(pre_amplification, int | float):
+        return pre_amplification
+
+    elif isinstance(pre_amplification, dict):
+        # Perform channel mismatch check after processing all gains
+        if geometry.channels is None:
+            raise ValueError("Missing parameter '.channels' in Geometry.")
+
+        defined_channels = set(geometry.channels.readout_position.keys())
+        input_channels = set(pre_amplification)
+        if defined_channels != input_channels:
+            raise ValueError(
+                "Mismatch between the defined channels in geometry and provided channel gains."
+            )
+
+        gain_map_2d = np.zeros(shape=geometry.shape, dtype=float)
+
+        for channel, gain in pre_amplification.items():
+            slice_y, slice_x = geometry.get_channel_coord(channel)
+            gain_map_2d[slice_y, slice_x] = gain
+
+        return gain_map_2d
+
+    else:
+        raise TypeError(
+            "'pre_amplification' must be a number, a dictionary of numbers or None"
+        )
 
 
 class Characteristics:
@@ -31,22 +133,22 @@ class Characteristics:
     ----------
     quantum_efficiency : float, optional
         Quantum efficiency.
-    charge_to_volt_conversion : float, optional
+    charge_to_volt : ChargeToVoltSettings, optional
         Sensitivity of charge readout. Unit: V/e-
     pre_amplification : float, optional
         Gain of pre-amplifier. Unit: V/V
     full_well_capacity : float, optional
         Full well capacity. Unit: e-
-    adc_voltage_range : tuple of floats, optional
-        ADC voltage range. Unit: V
     adc_bit_resolution : int, optional
         ADC bit resolution.
+    adc_voltage_range : tuple of floats, optional
+        ADC voltage range. Unit: V
     """
 
     def __init__(
         self,
         quantum_efficiency: float | None = None,  # unit: NA
-        charge_to_volt_conversion: float | None = None,  # unit: volt/electron
+        charge_to_volt: ChargeToVoltSettings | None = None,  # unit: volt/electron
         pre_amplification: float | dict[str, float] | None = None,  # unit: V/V
         full_well_capacity: float | None = None,  # unit: electron
         adc_bit_resolution: int | None = None,
@@ -55,17 +157,7 @@ class Characteristics:
         if quantum_efficiency is not None and not (0.0 <= quantum_efficiency <= 1.0):
             raise ValueError("'quantum_efficiency' must be between 0.0 and 1.0.")
 
-        if charge_to_volt_conversion is not None and not (
-            0.0 <= charge_to_volt_conversion <= 100.0
-        ):
-            raise ValueError(
-                "'charge_to_volt_conversion' must be between 0.0 and 100.0."
-            )
-
-        if isinstance(pre_amplification, float) and not (
-            0.0 <= pre_amplification <= 10_000.0
-        ):
-            raise ValueError("'pre_amplification' must be between 0.0 and 10000.0.")
+        validate_pre_amplification(pre_amplification)
 
         if full_well_capacity is not None and not (0.0 <= full_well_capacity <= 1.0e7):
             raise ValueError("'full_well_capacity' must be between 0 and 1e7.")
@@ -81,12 +173,12 @@ class Characteristics:
                 raise ValueError("Voltage range must have length of 2.")
 
         self._quantum_efficiency: float | None = quantum_efficiency
-        self._charge_to_volt_conversion: float | None = charge_to_volt_conversion
+        self._charge_to_volt: ChargeToVoltSettings | None = charge_to_volt
 
-        # TODO: This variable is available in class 'Characteristics' and 'APDCharacteristics'
-        #       Refactor this
-        self._channels_pre_amplification: float | np.ndarray | None = None
         self._pre_amplification: float | dict[str, float] | None = pre_amplification
+        self._pre_amplification_map: float | np.ndarray | None = (
+            None  # Geometry is not yet defined at this stage
+        )
 
         self._full_well_capacity: float | None = full_well_capacity
 
@@ -109,54 +201,18 @@ class Characteristics:
         return (
             type(self) is type(other)
             and self._quantum_efficiency == other._quantum_efficiency
-            and self._charge_to_volt_conversion == other._charge_to_volt_conversion
+            and self._charge_to_volt == other._charge_to_volt
             and self._pre_amplification == other._pre_amplification
             and self._full_well_capacity == other._full_well_capacity
             and self._adc_voltage_range == other._adc_voltage_range
             and self._adc_bit_resolution == other._adc_bit_resolution
         )
 
-    # TODO: This method exists in class 'Characteristics' and 'APDCharacteristics'
-    #       Refactor this
-    def _build_channels_gain(
-        self, value: int | float | dict[str, float] | None
-    ) -> None:
-        if value is None:
-            self._channels_pre_amplification = None
-
-        elif isinstance(value, float | int):
-            self._channels_pre_amplification = value
-
-        elif isinstance(value, dict):
-            if self._geometry is None:
-                raise ValueError(
-                    "Geometry must be initialized before setting channel gains."
-                )
-
-            if self._geometry.channels is None:
-                raise ValueError("Missing parameter '.channels' in Geometry.")
-
-            value_2d = np.zeros(shape=self._geometry.shape, dtype=float)
-
-            for channel, gain in value.items():
-                slice_y, slice_x = self._geometry.get_channel_coord(channel)
-                value_2d[slice_y, slice_x] = gain
-
-            self._channels_pre_amplification = value_2d
-
-            # Perform channel mismatch check after processing all gains
-            defined_channels = set(self._geometry.channels.readout_position.keys())
-            input_channels = set(value.keys())
-            if defined_channels != input_channels:
-                raise ValueError(
-                    "Mismatch between the defined channels in geometry and provided channel gains."
-                )
-
-    # TODO: This method is similar in 'APDCharacteristics and 'Characteristics'
-    #       Refactor these methods
     def initialize(self, geometry: "Geometry"):
         self._geometry = geometry
-        self._build_channels_gain(value=self._pre_amplification)
+        self._pre_amplification_map = to_pre_amplification_map(
+            self._pre_amplification, geometry=self._geometry
+        )
 
     @property
     def quantum_efficiency(self) -> float:
@@ -181,9 +237,9 @@ class Characteristics:
         self._quantum_efficiency = value
 
     @property
-    def charge_to_volt_conversion(self) -> float:
+    def charge_to_volt_conversion(self) -> float | np.ndarray:
         """Get charge to volt conversion parameter."""
-        if self._charge_to_volt_conversion is None:
+        if self._charge_to_volt is None:
             raise ValueError(
                 get_uninitialized_error(
                     name="charge_to_volt_conversion",
@@ -191,20 +247,19 @@ class Characteristics:
                 )
             )
 
-        return self._charge_to_volt_conversion
+        return self._charge_to_volt.value
 
     @charge_to_volt_conversion.setter
-    def charge_to_volt_conversion(self, value: float) -> None:
+    def charge_to_volt_conversion(self, value: float | np.ndarray) -> None:
         """Set charge to volt conversion parameter."""
-        if not (0.0 <= value <= 100.0):
-            raise ValueError(
-                "'charge_to_volt_conversion' must be between 0.0 and 100.0."
-            )
-        self._charge_to_volt_conversion = value
+        if self._charge_to_volt is None:
+            self._charge_to_volt = ChargeToVoltSettings(value=value)
+        else:
+            self._charge_to_volt.value = value
 
     @property
-    def channels_pre_amplification(self) -> float | np.ndarray:
-        if self._channels_pre_amplification is None:
+    def pre_amplification_map(self) -> float | np.ndarray:
+        if self._pre_amplification_map is None:
             raise ValueError(
                 get_uninitialized_error(
                     name="channels_gain",
@@ -212,7 +267,7 @@ class Characteristics:
                 )
             )
 
-        return self._channels_pre_amplification
+        return self._pre_amplification_map
 
     @property
     def pre_amplification(self) -> float | dict[str, float]:
@@ -229,23 +284,17 @@ class Characteristics:
 
     @pre_amplification.setter
     def pre_amplification(self, value: float | dict[str, float]) -> None:
-        """Set voltage pre-amplification gain.."""
-        if isinstance(value, float):
-            if not (0.0 <= value <= 10_000.0):
-                raise ValueError("'pre_amplification' must be between 0.0 and 10000.0.")
+        """Set voltage pre-amplification gain."""
+        if self._geometry is None:
+            raise RuntimeError(
+                "There is no 'Geometry' defined ! Please run method '.initialize(...)'"
+            )
 
-            self._build_channels_gain(value=value)
-            self._pre_amplification = value
+        validate_pre_amplification(value)
+        pre_amplification_map = to_pre_amplification_map(value, geometry=self._geometry)
 
-        elif isinstance(value, dict):
-            for channel, pre_ampli in value.items():
-                if not (0.0 <= pre_ampli <= 10_000.0):
-                    raise ValueError(
-                        f"'pre_amplification' must be between 0.0 and 10000.0. for {channel=}"
-                    )
-
-            self._build_channels_gain(value=value)
-            self._pre_amplification = value
+        self._pre_amplification = value
+        self._pre_amplification_map = pre_amplification_map
 
     @property
     def adc_bit_resolution(self) -> int:
@@ -313,7 +362,7 @@ class Characteristics:
         # TODO: Is it correct ?
         gain: Quantity = (
             self.quantum_efficiency
-            * Quantity(self.channels_pre_amplification, unit="V/V")
+            * Quantity(self.pre_amplification_map, unit="V/V")
             * Quantity(2**self.adc_bit_resolution, unit="adu")
         ) / (
             np.max(Quantity(self.adc_voltage_range, unit="V"))
@@ -338,7 +387,9 @@ class Characteristics:
         """Get the attributes of this instance as a `dict`."""
         return {
             "quantum_efficiency": self._quantum_efficiency,
-            "charge_to_volt_conversion": self._charge_to_volt_conversion,
+            "charge_to_volt": (
+                self._charge_to_volt.to_dict() if self._charge_to_volt else None
+            ),
             "pre_amplification": self._pre_amplification,
             "full_well_capacity": self._full_well_capacity,
             "adc_bit_resolution": self._adc_bit_resolution,
@@ -346,10 +397,13 @@ class Characteristics:
         }
 
     @classmethod
-    def from_dict(cls, dct: Mapping):
+    def from_dict(cls, dct: Mapping | None) -> Self:
         """Create a new instance from a `dict`."""
         # Late import to speedup start-up time
         from toolz import dicttoolz
+
+        if dct is None:
+            dct = {}
 
         # TODO: This is a simplistic implementation. Improve this.
         # Extract param 'adc_voltage_range'
@@ -362,14 +416,47 @@ class Characteristics:
             adc_voltage_min, adc_voltage_max = tuple(param)
             adc_voltage_range = adc_voltage_min, adc_voltage_max
 
-        return cls(adc_voltage_range=adc_voltage_range, **new_dct)
+        charge_to_volt: ChargeToVoltSettings | None = None
+        if "charge_to_volt_conversion" in dct:
+            warnings.warn(
+                "Parameter 'charge_to_volt_conversion' is deprecated. "
+                "Use parameter 'charge_to_volt'",
+                DeprecationWarning,
+                stacklevel=1,
+            )
+
+            if "charge_to_volt" in dct:
+                raise ValueError(
+                    "Cannot provide parameter 'charge_to_volt_conversion' and 'charge_to_volt'"
+                )
+
+            charge_to_volt_value = dct["charge_to_volt_conversion"]
+            charge_to_volt = ChargeToVoltSettings(value=charge_to_volt_value)
+
+            new_dct = dicttoolz.dissoc(dct, "charge_to_volt_conversion")
+
+            return cls(charge_to_volt=charge_to_volt, **new_dct)
+
+        charge_to_volt_dct: Mapping | None = new_dct.get("charge_to_volt")
+        new_dct = dicttoolz.dissoc(new_dct, "charge_to_volt")
+
+        if charge_to_volt_dct:
+            charge_to_volt = ChargeToVoltSettings.from_dict(charge_to_volt_dct)
+
+        return cls(
+            charge_to_volt=charge_to_volt,
+            adc_voltage_range=adc_voltage_range,
+            **new_dct,
+        )
 
     def dump(
         self,
     ) -> dict[str, int | float | str | dict[str, float] | list[float] | None]:
         return {
             "quantum_efficiency": self._quantum_efficiency,
-            "charge_to_volt_conversion": self._charge_to_volt_conversion,
+            "charge_to_volt_conversion": (
+                self._charge_to_volt.to_dict() if self._charge_to_volt else None
+            ),
             "pre_amplification": self._pre_amplification,
             "full_well_capacity": self._full_well_capacity,
             "adc_bit_resolution": self._adc_bit_resolution,