+++ pyxel/models/charge_generation/apd_gain.py
@@ -30,11 +30,11 @@ def apd_gain(detector: APD) -> None:
     if detector.charge.frame_empty():
         array_copy = detector.charge.array.copy()
         detector.charge.empty()
-        array_copy *= detector.characteristics.avalanche_gain
+        array_copy *= detector.characteristics.avalanche_settings.avalanche_gain
         detector.charge.add_charge_array(array=array_copy)
 
     else:
         frame_copy = detector.charge.frame.copy()
         detector.charge.empty()
-        frame_copy.number *= detector.characteristics.avalanche_gain
+        frame_copy.number *= detector.characteristics.avalanche_settings.avalanche_gain
         detector.charge.add_charge_dataframe(new_charges=frame_copy)